Refactor send assets action handling and move utility to new file
